backend/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "college_advisory"
version = "0.1.3"
dependencies = [
    "Flask",
    "python-dotenv",
    "openai",
    # 其他依赖项
]
